Add files via upload
The client init function now works between the collector and the client. Added a configuration script to customize the client, collector, and server. Added functionality to the server to log the check ins and command outputs of the NightShift client. The logging output can now be picked up and set to something like an ELK Stack.