Add files via upload
The utility script is a combination of the dga and cipher scripts. Updated the client, server and init scripts to reflect the new import. Also combined the configuration and four o four script into one so it has two modes to write the fof cmd and configure the client and server.